docker-requirements.txt
# Core Dependencies
slack-sdk==3.23.0
python-dotenv==1.0.0
requests==2.31.0

# Notebook Execution
papermill==2.5.0
nbformat==5.9.0
ipykernel==6.25.0

# Data Processing
pandas==2.1.0
numpy==1.24.0

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# Database (Presto)
prestodb==0.8.0

# Additional notebook dependencies
matplotlib==3.8.0
scipy==1.11.0
scikit-learn==1.3.0
looker-sdk==23.20.0

# Logging
coloredlogs==15.0

# Utilities
pytz==2023.3
dateutil==2.8.2
